requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
faiss-cpu
jinja2==3.1.4
python-multipart==0.0.9
numpy
scikit-learn==1.8.0
